git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/locate-loop-and-live-variables@47751 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/bdd/BDDCompressExpressionVisitor.java
@@ -160,7 +160,7 @@ public class BDDCompressExpressionVisitor extends DefaultCExpressionVisitor<Regi
   @Override
   public Region[] visit(CIdExpression idExp) {
     if (idExp.getDeclaration() instanceof CEnumerator enumerator) {
-      return intToRegions.get(BigInteger.valueOf(enumerator.getValue()));
+      return intToRegions.get(enumerator.getValue());
     }
     return predMgr.createPredicate(
         idExp.getDeclaration().getQualifiedName(),
